Keep concurrency gate analysis and tests together
Amp-Thread-ID: https://ampcode.com/threads/T-01a08946-5cf3-77e8-806c-5d31f591b6c7
